Refactor authentication alerts
Refactored the authentication alerts to disappear after a period of
time. Created new alert divs instead of reusing one div. Utilized a
flexbox to allow multiple alerts to stack up. Moved the welcome message
to the center of the page to be more prominent. Updated the nav logo
to be consistent with theme.